requirements.txt
# requirements.txt
ollama
click
requests
beautifulsoup4
html2text
pyyaml